pyproject.toml
[project]
name = "youtube-study-buddy"
version = "0.1.0"
description = "YouTube transcript to study notes converter with AI integration"
authors = [
    {name = "Justin Robinson", email = "justin.g.robinson@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "youtube-transcript-api>=0.6.2",
    "anthropic>=0.25.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "urllib3>=2.0.0",
    "jupyter>=1.0.0",
    "notebook>=7.0.0",
    "ipykernel>=6.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "scikit-learn>=1.3.0",
    "sentence-transformers>=2.2.0",
    "scipy>=1.11.0",
    "streamlit>=1.0.0",
    "pytest>=8.4.2",
    "PySocks>=1.7.1",
    "stem>=1.8.0",
    "yt-dlp>=2025.9.26",
    "fuzzywuzzy>=0.18.0",
    "python-Levenshtein>=0.21.0",
    "weasyprint>=60.0",
    "markdown2>=2.4.0",
    "loguru>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[project.scripts]
youtube-study-buddy = "yt_study_buddy.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/yt_study_buddy"]
